requirements.txt
vietocr
torch==2.3.1
torchvision==0.2.1
opencv-python==4.10.0.82
scikit-image==0.21.0
scipy==1.10.1
pdfplumber==0.11.2
pypdf==4.3.1
python-docx==1.1.2
pdf2image
transformers==4.42.4
sentencepiece==0.2.0
underthesea==6.8.4